+++ geotrek/trekking/models.py
@@ -1,11 +1,9 @@
 import os
 import logging
 import re
-import json
 
 from django.conf import settings
 from django.contrib.gis.db import models
-from django.core.urlresolvers import reverse
 from django.utils.translation import ugettext_lazy as _
 
 from easy_thumbnails.alias import aliases
@@ -94,13 +92,10 @@ class Trek(PicturesMixin, PublishableMixin, MapEntityMixin, Topology):
     def __unicode__(self):
         return self.name
 
-    @property
-    @models.permalink
-    def elevation_area_url(self):
-        return ('trekking:trek_elevation_area', [str(self.pk)])
-
     @models.permalink
     def get_document_public_url(self):
+        """ Override ``geotrek.common.mixins.PublishableMixin``
+        """
         return ('trekking:trek_document_public', [str(self.pk)])
 
     @property
@@ -124,71 +119,6 @@ class Trek(PicturesMixin, PublishableMixin, MapEntityMixin, Topology):
         return POIType.objects.filter(pk__in=set(pks))
 
     @property
-    def serializable_relationships(self):
-        return [{
-                'has_common_departure': rel.has_common_departure,
-                'has_common_edge': rel.has_common_edge,
-                'is_circuit_step': rel.is_circuit_step,
-                'trek': {
-                    'pk': rel.trek_b.pk,
-                    'slug': rel.trek_b.slug,
-                    'name': rel.trek_b.name,
-                    'url': reverse('trekking:trek_json_detail', args=(rel.trek_b.pk,)),
-                },
-                'published': rel.trek_b.published} for rel in self.relationships]
-
-    @property
-    def serializable_cities(self):
-        return [{'code': city.code,
-                 'name': city.name} for city in self.cities]
-
-    @property
-    def serializable_networks(self):
-        return [{'id': network.id,
-                 'pictogram': network.serializable_pictogram,
-                 'name': network.network} for network in self.networks.all()]
-
-    @property
-    def serializable_difficulty(self):
-        if not self.difficulty:
-            return None
-        return {'id': self.difficulty.pk,
-                'pictogram': self.difficulty.serializable_pictogram,
-                'label': self.difficulty.difficulty}
-
-    @property
-    def serializable_themes(self):
-        return [{'id': t.pk,
-                 'pictogram': t.serializable_pictogram,
-                 'label': t.label} for t in self.themes.all()]
-
-    @property
-    def serializable_usages(self):
-        return [{'id': u.pk,
-                 'pictogram': u.serializable_pictogram,
-                 'label': u.usage} for u in self.usages.all()]
-
-    @property
-    def serializable_districts(self):
-        return [{'id': d.pk,
-                 'name': d.name} for d in self.districts]
-
-    @property
-    def serializable_route(self):
-        if not self.route:
-            return None
-        return {'id': self.route.pk,
-                'pictogram': self.route.serializable_pictogram,
-                'label': self.route.route}
-
-    @property
-    def serializable_web_links(self):
-        return [{'id': w.pk,
-                 'name': w.name,
-                 'category': w.serializable_category,
-                 'url': w.url} for w in self.web_links.all()]
-
-    @property
     def information_desk(self):
         """Retrocompatibily method for Geotrek-rando.
         """
@@ -196,27 +126,6 @@ class Trek(PicturesMixin, PublishableMixin, MapEntityMixin, Topology):
             return self.information_desks.first()
         except (ValueError, InformationDesk.DoesNotExist):
             return None
-
-    @property
-    def serializable_information_desk(self):
-        return self.information_desk.__json__() if self.information_desk else None
-
-    @property
-    def serializable_information_desks(self):
-        return [d.__json__() for d in self.information_desks.all()]
-
-    @property
-    def serializable_parking_location(self):
-        if not self.parking_location:
-            return None
-        return self.parking_location.transform(settings.API_SRID, clone=True).coords
-
-    @property
-    def serializable_points_reference(self):
-        if not self.points_reference:
-            return None
-        geojson = self.points_reference.transform(settings.API_SRID, clone=True).geojson
-        return json.loads(geojson)
 
     @property
     def length_kilometer(self):
@@ -441,13 +350,6 @@ class WebLink(models.Model):
     def get_add_url(cls):
         return ('trekking:weblink_add', )
 
-    @property
-    def serializable_category(self):
-        if not self.category:
-            return None
-        return {'label': self.category.label,
-                'pictogram': self.category.serializable_pictogram}
-
 
 class WebLinkCategory(PictogramMixin):
 
@@ -534,23 +436,10 @@ class InformationDesk(models.Model):
     def __unicode__(self):
         return self.name
 
-    def __json__(self):
-        return {
-            'name': self.name,
-            'description': self.description,
-            'phone': self.phone,
-            'email': self.email,
-            'website': self.website,
-            'photo_url': self.photo_url,
-            'street': self.street,
-            'postal_code': self.postal_code,
-            'municipality': self.municipality,
-            'latitude': self.latitude,
-            'longitude': self.longitude,
-        }
-
     @property
     def description_strip(self):
+        """Used in trek public template.
+        """
         nobr = re.compile(r'(\s*<br.*?>)+\s*', re.I)
         newlines = nobr.sub("\n", self.description)
         return smart_plain_text(newlines)
@@ -614,7 +503,7 @@ class POI(PicturesMixin, PublishableMixin, MapEntityMixin, Topology):
     @property
     def serializable_type(self):
         return {'label': self.type.label,
-                'pictogram': self.type.serializable_pictogram}
+                'pictogram': self.type.get_pictogram_url()}
 
     @classmethod
     def path_pois(cls, path):